refactor: Converted all references to Double to Decimals used by location

+++ WIKILocationSelector/Data/LocationData.swift
@@ -10,8 +10,8 @@ import Foundation
 struct LocationData: Equatable, Decodable {
         
     let name: String?
-    let latitude: Double
-    let longitude: Double
+    let latitude: Decimal
+    let longitude: Decimal
     enum CodingKeys: String, CodingKey {
         case name
         case latitude = "lat"
